+++ src/refactoring/components/admin/product/EditProductForm.tsx
@@ -1,0 +1,177 @@
+import { useState } from "react";
+import { Discount, Product } from "../../../../types";
+
+interface EditProductFormProps {
+  product: Product;
+  products: Product[];
+  editingProduct: Product;
+  setEditingProduct: React.Dispatch<React.SetStateAction<Product | null>>;
+  onProductUpdate: (updatedProduct: Product) => void;
+}
+
+const EditProductForm = ({
+  product,
+  products,
+  editingProduct,
+  setEditingProduct,
+  onProductUpdate,
+}: EditProductFormProps) => {
+  const [newDiscount, setNewDiscount] = useState<Discount>({
+    quantity: 0,
+    rate: 0,
+  });
+
+  const handleProductNameUpdate = (productId: string, newName: string) => {
+    if (editingProduct && editingProduct.id === productId) {
+      const updatedProduct = { ...editingProduct, name: newName };
+      setEditingProduct(updatedProduct);
+    }
+  };
+
+  const handlePriceUpdate = (productId: string, newPrice: number) => {
+    if (editingProduct && editingProduct.id === productId) {
+      const updatedProduct = { ...editingProduct, price: newPrice };
+      setEditingProduct(updatedProduct);
+    }
+  };
+
+  const handleStockUpdate = (productId: string, newStock: number) => {
+    const updatedProduct = products.find((p) => p.id === productId);
+    if (updatedProduct) {
+      const newProduct = { ...updatedProduct, stock: newStock };
+      onProductUpdate(newProduct);
+      setEditingProduct(newProduct);
+    }
+  };
+
+  const handleEditComplete = () => {
+    if (editingProduct) {
+      onProductUpdate(editingProduct);
+      setEditingProduct(null);
+    }
+  };
+
+  const handleAddDiscount = (productId: string) => {
+    const updatedProduct = products.find((p) => p.id === productId);
+    if (updatedProduct && editingProduct) {
+      const newProduct = {
+        ...updatedProduct,
+        discounts: [...updatedProduct.discounts, newDiscount],
+      };
+      onProductUpdate(newProduct);
+      setEditingProduct(newProduct);
+      setNewDiscount({ quantity: 0, rate: 0 });
+    }
+  };
+
+  const handleRemoveDiscount = (productId: string, index: number) => {
+    const updatedProduct = products.find((p) => p.id === productId);
+    if (updatedProduct) {
+      const newProduct = {
+        ...updatedProduct,
+        discounts: updatedProduct.discounts.filter((_, i) => i !== index),
+      };
+      onProductUpdate(newProduct);
+      setEditingProduct(newProduct);
+    }
+  };
+  return (
+    // 제품 수정 폼 컴포넌트
+    <div>
+      {/* 제품명 입력 컴포넌트 */}
+      <div className="mb-4">
+        <label className="block mb-1">상품명: </label>
+        <input
+          type="text"
+          value={editingProduct.name}
+          onChange={(e) => handleProductNameUpdate(product.id, e.target.value)}
+          className="w-full p-2 border rounded"
+        />
+      </div>
+      {/* 가격 입력 컴포넌트 */}
+      <div className="mb-4">
+        <label className="block mb-1">가격: </label>
+        <input
+          type="number"
+          value={editingProduct.price}
+          onChange={(e) =>
+            handlePriceUpdate(product.id, parseInt(e.target.value))
+          }
+          className="w-full p-2 border rounded"
+        />
+      </div>
+      {/* 재고 입력 컴포넌트 */}
+      <div className="mb-4">
+        <label className="block mb-1">재고: </label>
+        <input
+          type="number"
+          value={editingProduct.stock}
+          onChange={(e) =>
+            handleStockUpdate(product.id, parseInt(e.target.value))
+          }
+          className="w-full p-2 border rounded"
+        />
+      </div>
+      {/* 할인 정보 컴포넌트 */}
+      <div>
+        <h4 className="text-lg font-semibold mb-2">할인 정보</h4>
+        {/* 할인 목록 컴포넌트 */}
+        {editingProduct.discounts.map((discount, index) => (
+          <div key={index} className="flex justify-between items-center mb-2">
+            <span>
+              {discount.quantity}개 이상 구매 시 {discount.rate * 100}% 할인
+            </span>
+            <button
+              onClick={() => handleRemoveDiscount(product.id, index)}
+              className="bg-red-500 text-white px-2 py-1 rounded hover:bg-red-600"
+            >
+              삭제
+            </button>
+          </div>
+        ))}
+        {/* 새 할인 추가 컴포넌트 */}
+        <div className="flex space-x-2">
+          <input
+            type="number"
+            placeholder="수량"
+            value={newDiscount.quantity}
+            onChange={(e) =>
+              setNewDiscount({
+                ...newDiscount,
+                quantity: parseInt(e.target.value),
+              })
+            }
+            className="w-1/3 p-2 border rounded"
+          />
+          <input
+            type="number"
+            placeholder="할인율 (%)"
+            value={newDiscount.rate * 100}
+            onChange={(e) =>
+              setNewDiscount({
+                ...newDiscount,
+                rate: parseInt(e.target.value) / 100,
+              })
+            }
+            className="w-1/3 p-2 border rounded"
+          />
+          <button
+            onClick={() => handleAddDiscount(product.id)}
+            className="w-1/3 bg-blue-500 text-white p-2 rounded hover:bg-blue-600"
+          >
+            할인 추가
+          </button>
+        </div>
+      </div>
+      {/* 수정 완료 버튼 컴포넌트 */}
+      <button
+        onClick={handleEditComplete}
+        className="bg-green-500 text-white px-2 py-1 rounded hover:bg-green-600 mt-2"
+      >
+        수정 완료
+      </button>
+    </div>
+  );
+};
+
+export default EditProductForm;
